This change is a huge improvement:

```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              343266        296104        -13.74%
BenchmarkDeducedNested-12                              900868        765058        -15.08%
BenchmarkDeducedNestedAcrossVersion-12                 1004928       799466        -20.45%
BenchmarkLeafConflictAcrossVersion-12                  349167        279827        -19.86%
BenchmarkMultipleApplierRecursiveRealConversion-12     2231486       2149787       -3.66%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              1264           1098           -13.13%
BenchmarkDeducedNested-12                              3341           2965           -11.25%
BenchmarkDeducedNestedAcrossVersion-12                 3602           3112           -13.60%
BenchmarkLeafConflictAcrossVersion-12                  1296           1078           -16.82%
BenchmarkMultipleApplierRecursiveRealConversion-12     6841           6476           -5.34%
```

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1605549       1355981       -15.54%
BenchmarkUpdateNewObject-12     1752232       1565341       -10.67%
BenchmarkRepeatedUpdate-12      521986        443528        -15.03%
BenchmarkSetToFields-12         5156          5154          -0.04%
BenchmarkFieldsToSet-12         9908          10050         +1.43%
BenchmarkYAMLToTyped-12         54713         56860         +3.92%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      5045           4264           -15.48%
BenchmarkUpdateNewObject-12     5839           4850           -16.94%
BenchmarkRepeatedUpdate-12      1592           1255           -21.17%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             82             +0.00%
BenchmarkYAMLToTyped-12         90             90             +0.00%
```